🚚rename: add HybridDialog folder
Related to: #86
